Make Handler work with new dispatcher and reorganize handler module
