--- src/core/handler/mod.rs
@@ -1,5 +1,0 @@
-pub(in crate::core) mod base;
-mod command;
-mod predicate;
-
-pub use self::{base::*, command::*, predicate::*};
