+++ src/core/mod.rs
@@ -1,6 +1,8 @@
+mod app;
 mod context;
 mod convert;
 mod dispatcher;
 mod handler;
+mod predicate;
 
-pub use self::{context::*, convert::*, dispatcher::*, handler::*};
+pub use self::{app::*, context::*, convert::*, dispatcher::*, handler::*, predicate::*};
