+++ src/core/predicate/mod.rs
@@ -1,0 +1,6 @@
+mod base;
+mod command;
+mod ext;
+mod result;
+
+pub use self::{base::*, command::*, ext::*, result::*};
